fix: isolate triage payload helpers
